refac: 매칭 관리 탭 부분 기능 controller 통합
