+++ src/main/kotlin/org/appjam/smashing/domain/game/controller/GameController.kt
@@ -9,9 +9,12 @@ import org.appjam.smashing.domain.game.dto.request.GameResultSubmitRequest
 import org.appjam.smashing.domain.game.dto.response.GameResultConfirmResponse
 import org.appjam.smashing.domain.game.dto.response.GameResultSubmissionDetailResponse
 import org.appjam.smashing.domain.game.dto.response.GameResultSubmitResponse
+import org.appjam.smashing.domain.game.dto.response.PendingResultAcceptedGameSummaryResponse
 import org.appjam.smashing.domain.game.service.GameService
 import org.appjam.smashing.global.auth.security.data.CustomUserDetails
 import org.appjam.smashing.global.common.dto.ApiResponse
+import org.appjam.smashing.global.common.dto.CommonCursorRequest
+import org.appjam.smashing.global.common.dto.CursorResponse
 import org.springframework.http.ResponseEntity
 import org.springframework.security.core.annotation.AuthenticationPrincipal
 import org.springframework.web.bind.annotation.*
@@ -122,6 +125,47 @@ class GameController(
         val response = gameService.getSubmissionDetail(
             gameId = gameId,
             submissionId = submissionId,
+        )
+
+        return ApiResponse.success(response)
+    }
+
+    @Operation(
+        summary = "결과 확정 전(수락된) 게임 목록 조회 API",
+        description = """
+            매칭 수락(ACCEPTED)되어 게임이 자동 생성된 이후, 아직 결과가 확정(RESULT_CONFIRMED)되지 않은 게임 목록을 조회합니다.
+
+            - 조회 대상(resultStatus):
+              - PENDING_RESULT / WAITING_CONFIRMATION / RESULT_REJECTED (CONFIRMED, CANCELED 제외)
+
+            - 기준 스포츠:
+              - 로그인 유저의 활성 프로필(activeUserSportProfileId)의 sport
+
+            - 정렬 : order 파라미터 기반 정렬
+              - LATEST / latest / Latest 등 대소문자 무관
+              - OLDEST / oldest / Oldest 등 대소문자 무관
+              - 기본값(default) : order 미지정 또는 잘못된 값 → LATEST 적용
+              - 정렬 기준 :
+                - LATEST : gameId (TSID) DESC
+                - OLDEST : gameId (TSID) ASC
+
+            - 페이징:
+              - snapshotAt 이전 데이터만 조회(스냅샷 고정)
+              - cursor = gameId (마지막 gameId)
+
+            - 잠금(결과 제출 가능 시각) 계산:
+              - 목록은 전체(스냅샷 이전) 조회
+              - 잠금 시간 계산만 '오늘 기준 정책'으로 계산해 내려줌
+        """
+    )
+    @GetMapping("/pending-results")
+    fun getPendingResultAcceptedGames(
+        @AuthenticationPrincipal principal: CustomUserDetails,
+        @Valid request: CommonCursorRequest,
+    ): ResponseEntity<ApiResponse<CursorResponse<PendingResultAcceptedGameSummaryResponse>>> {
+        val response = gameService.getPendingResultAcceptedGames(
+            userId = principal.username,
+            request = request,
         )
 
         return ApiResponse.success(response)
